requirements.txt
scikit-learn
pillow
streamlit
numpy
tensorflow
opencv-python-headless